requirements.txt
pandas
boto3
sagemaker
awscli
lightning=2.1.2
torch-tb-profiler==0.4.3